(#741) Fix failing favicon retrieval
